backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
python-decouple>=3.8
django-encrypted-model-fields>=0.6
django-ratelimit>=4.1
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<23.0
dj-database-url>=2.2.0
whitenoise[brotli]>=6.7.
Pillow